requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.7.0
cffi==1.15.1
channels==3.0.5
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
mysqlclient==2.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2022.2.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
zope.interface==5.4.0
